+++ src/term.h
@@ -448,11 +448,12 @@ struct term {
   unsigned int app_control;
   bool app_cursor_keys;
   bool app_keypad;
-  bool app_wheel;
   bool auto_repeat;
   bool bell_taskbar; // xterm: bellIsUrgent; switchable with CSI ? 1042 h/l
   bool bell_popup;   // xterm: popOnBell;    switchable with CSI ? 1043 h/l
-  bool wheel_reporting;
+  bool wheel_reporting_xterm; // xterm: alternateScroll
+  bool wheel_reporting;       // similar, but default true
+  bool app_wheel;             // format for wheel_reporting
   int  modify_other_keys;
   bool newline_mode;
   bool report_focus;